added appointment detail page

+++ lib/models/users/user_detail_model.dart
@@ -69,7 +69,7 @@ class UserDetailModel {
 }
 
 class PayrollCalculationsResponse {
-  int? salary;
+  double? salary;
   String? percentOrFixed;
 
   PayrollCalculationsResponse({this.salary, this.percentOrFixed});
